apply style changes to the Compressor
